requirements.txt
aiofiles==24.1.0
aiogram==3.17.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
certifi==2025.1.31
colorama==0.4.6
distro==1.9.0
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
magic-filter==1.0.12
multidict==6.1.0
openai==1.61.0
propcache==0.2.1
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
setuptools==75.8.0
sniffio==1.3.1
tqdm==4.67.1
typing_extensions==4.12.2
wheel==0.44.0
yarl==1.18.3
